Add standings subcommand!
Added Standing, DivisionStandings and DivisionStandingsProvider classes.
A Standing represents one line of a group of standings, basically the
stats for one team. An instance of DivisionStandings represents the
group of Standing instances for a given division at a given time. And of
course the DivisionStandingsProvider is what you can use to look up
these standings by division name.

Note: I decided to rename all instances of "league" in the code and docs
with "division" because it seems more accurate. PVA is a league. Coed A
is a division.