Use schedules (plural) subcommand for all or single team schedules
I kept typing in pva schedules 'team' and it would give me an error
about the wrong number of arguments. Bad error handling aside, I think
it's reasonable to allow the user to type either schedule or schedules.
So I made schedules the actual command which either returns all the
schedules for all of your teams or accepts a team name argument. I also
aliased that command to schedule, so you can use either term and it will
work for both use cases.