Format byes differently than normal matches for easy detection
Since byes are listed like normal matches on the website there is a
date, time, opposing team and location, so it can be formatted just like
a match with the location being "BYE", but it really doesn't make sense
to list all that information. All you need to know is that your team has
a BYE for that date.

So, a bye is now listed with the date and hyphens where the time would
be, and just the word BYE for the match description. This makes it
really stand out so you don't accidentally think you have a match that
day.

Also stripped extra spaces off of the match fields as they are being
extracted from the html.