Add SetResult, MatchResults classes. Methods to Match to work with them.
Also setup some initial specs for that functionality. This is all in
preparation for implementing the scores subcommand. It will display the
scores in a nice format tailored to the team's perspective, and will add
the result (W, L, or T) and the match points earned 0.0 - 4.5.

From a Match instance you can request the match results from a given
team's perspective. So, if you have a match between Court Jesters and Mo
Diggity, you can request the results from either team's perspective.

This is a WIP commit, where the specs are requiring this whole script,
which actually runs it, displaying the help output. This is less than
ideal. I need to extract the different classes into their own files and
require only the class(es) under test in each spec file, which should
also be broken out into multiple files.

Also, I have some comments and commented out code to get me started on
implementing the scores functionality in CLI. I might first extract the
classes and specs into their own files, then move on to implement the
rest of the scores subcommand functionality.