Handle empty scores and standings scenarios
If a team in your teams cache (say an indoor team) is not playing during
the current season (say summer season), and therefore there is no
standings for that team on the standings page, the provider needs to
handle that case cleanly, detect no standings and just display none.
Same thing with scores. If your team hasn't played any games yet during
the given season or if your team isn't playing at all that season, it
needs to detect that and display no results appropriately (instead of
crashing... haha).