Update URLs to use HTTPs. Bump version to 0.6.0.
Note that I intentionally deleted some history with a force push, which
is why the version seems to jump from 0.2.1 to 0.6.0. The missing
versions were misguided attempts to work around an issue that was
causing the results to not be properly filtered. I misdiagnosed the
reason for these errors as being caused by client side filtering of
results, but the real cause turned out to be just a change from http to
https, so when the program requested the http versions it was just
receiving a 301 redirect and getting the unfiltered version of the
requested page.

So, the quickest and easiest solution was to revert to the previous
version of the code that expected the filtering to be done on the server
side of the PVA website, and just update the URLs to use https, which is
what I ought to have done when I first saw the issue.